Move non-listener tests out of test_listeners
